Use
Outputting the probability of sentiment relative to the batch in which it was created renders confidence scores meaningless for comparison across datasets and for use in scoring. Replace this with a 1 so that this can be used as an absolute and/or make a relative probability within the batch as well.